requirements.txt
acfunsdk>=0.9.7
rich>=12.5
textual>=0.1.18
click>=8.1
Pillow>=9
kdtree>=0.16
poetry>=1.2
